requirements.txt
flask
Flask-SQLAlchemy
Flask-WTF
email-validator
werkzeug
scikit-learn
pandas
numpy
joblib
matplotlib
fpdf2
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
seaborn
